pyproject.toml
[project]
name = "assessorai-pdftomd"
version = "0.2.0"
description = "Convert legislative PDFs to Markdown using Gemini Vision API"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
]
